Cleanup test helpers from K8s
