+++ kubernetes/tests/integration/integration/high_availability/conftest.py
@@ -3,16 +3,15 @@
 # See LICENSE file for licensing details.
 
 import logging
+import os
+import subprocess
 from collections.abc import Generator
 
 import pytest
 from jubilant_backports import Juju
+from tenacity import RetryError, Retrying, stop_after_delay, wait_fixed
 
 from ...helpers_ha import get_app_leader
-from .high_availability_helpers import (
-    deploy_chaos_mesh,
-    destroy_chaos_mesh,
-)
 
 MYSQL_TEST_APP_NAME = "mysql-test-app"
 
@@ -43,3 +42,44 @@ def chaos_mesh(juju: Juju) -> Generator:
 
     logging.info("Destroying chaos mesh")
     destroy_chaos_mesh(juju.model)
+
+
+def deploy_chaos_mesh(namespace: str) -> None:
+    """Deploy chaos mesh to the provided namespace."""
+    env = os.environ
+    env["KUBECONFIG"] = os.path.expanduser("~/.kube/config")
+
+    subprocess.check_output(
+        f"tests/integration/integration/high_availability/scripts/deploy_chaos_mesh.sh {namespace}",
+        shell=True,
+        env=env,
+    )
+
+    try:
+        for attempt in Retrying(stop=stop_after_delay(5 * 60), wait=wait_fixed(10)):
+            with attempt:
+                output = subprocess.check_output(
+                    [
+                        "microk8s.kubectl",
+                        "get",
+                        "pods",
+                        f"--namespace={namespace}",
+                        "--selector=app.kubernetes.io/instance=chaos-mesh",
+                    ],
+                    env=env,
+                )
+                assert output.decode().count("Running") == 4, "Chaos Mesh not ready"
+    except RetryError:
+        raise Exception("Chaos Mesh pods not found") from None
+
+
+def destroy_chaos_mesh(namespace: str) -> None:
+    """Remove chaos mesh from the provided namespace."""
+    env = os.environ
+    env["KUBECONFIG"] = os.path.expanduser("~/.kube/config")
+
+    subprocess.check_output(
+        f"tests/integration/integration/high_availability/scripts/destroy_chaos_mesh.sh {namespace}",
+        shell=True,
+        env=env,
+    )
